Se desarrollo el resgitro de productos

+++ database/migrations/2019_08_03_223429_create_productos_table.php
@@ -19,7 +19,7 @@ class CreateProductosTable extends Migration
             $table->string('imagen2');
             $table->string('imagen3');
             $table->string('nombre');
-            $table->decimal('valor');
+            $table->integer('valor');
             $table->integer('cantidad');
             $table->integer('stock_minimo');
             $table->string('opcion_catologo');
